+++ src/main/java/com/prj1/controller/MemberController.java
@@ -3,6 +3,7 @@ package com.prj1.controller;
 import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -41,8 +42,10 @@ public class MemberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
         return "redirect:/member/list";
     }
 
@@ -75,5 +78,15 @@ public class MemberController {
     public String nickNameCheck(String nickName) {
         String message = service.nickCheck(nickName);
         return message;
+    }
+
+    @GetMapping("login")
+    public String loginForm() {
+        return "member/login";
+    }
+
+    @PostMapping("login")
+    public String login() {
+        return "redirect:board/";
     }
 }
